Fix windows method to take usize in place of u64

+++ core/io/windows.rs
@@ -117,7 +117,7 @@ impl File for WindowsFile {
         Ok(c)
     }
 
-    fn truncate(&self, len: u64, c: Completion) -> Result<()> {
+    fn truncate(&self, len: usize, c: Completion) -> Result<()> {
         let mut file = self.file.borrow_mut();
         file.set_len(len).map_err(LimboError::IOError)?;
         c.complete(0);
